database/requirements.txt
pymongo>=4.7
python-dotenv
PyJWT